Fix ordered fragement q validation
q is Numeric, not always Float.

+++ ruby/lib/trailer_vote/media_types/ordered_fragment.rb
@@ -15,7 +15,7 @@ module TrailerVote
             attribute :ordered_fragments do
               collection :_index, allow_empty: true do
                 attribute :href, Types::HttpUrl
-                attribute :q, Float
+                attribute :q, Numeric
               end
               not_strict
             end
